+++ chartlets.js/CHANGES.md
@@ -6,6 +6,9 @@
 	- `vite: ^7.1.11`
 	- `vitest: ^3.2.4`
 
+* Added icon support for `Button`, `IconButton` and `Tabs` components.
+ (#124).
+
 ## Version 0.1.7 (from 2025/12/03)
 
 * Updated dependencies